feat(explorer): add close button
